+++ db/migrate/20140218195245_add_collect_email_on_surveys_to_project.rb
@@ -1,0 +1,5 @@
+class AddCollectEmailOnSurveysToProject < ActiveRecord::Migration
+  def change
+    add_column :projects, :collect_email_on_surveys, :boolean, null: false, default: true
+  end
+end